Move parseArray to ArrayStream module
