+++ src/Streamly/Internal/Data/Stream/IsStream/Eliminate.hs
@@ -32,7 +32,6 @@ module Streamly.Internal.Data.Stream.IsStream.Eliminate
     , parseD
     , parse_
     , parseD_
-    , parseArrayD
 
     -- * Stream Deconstruction
     -- | foldr and foldl do not provide the remaining stream.  'uncons' is more
@@ -351,18 +350,6 @@ runSink = fold . toFold
 {-# INLINE_NORMAL parseD #-}
 parseD :: MonadThrow m => PRD.Parser m a b -> SerialT m a -> m b
 parseD p = D.parse p . toStreamD
-
--- | Parse an array stream using the supplied 'Parser'.
---
--- /Internal/
---
-{-# INLINE_NORMAL parseArrayD #-}
-parseArrayD ::
-       (MonadIO m, Storable a)
-    => PRD.Parser m a b
-    -> SerialT m (A.Array a)
-    -> m (b, SerialT m (A.Array a))
-parseArrayD p s = fmap fromStreamD <$> D.parseArray p (toStreamD s)
 
 -- | Parse a stream using the supplied ParserK 'PRK.Parser'.
 --
